Move schedules and sets to `lib.rs` and remove things from prelude
